refactor: split plugin page helpers
