+++ src/static/scripts/plugin_page.js
@@ -1,174 +1,18 @@
 (function () {
-  function validateAddToPlaylistAction(action) {
-    if (action !== "add_to_playlist") return true;
-    const instanceInput = document.getElementById("instance");
-    const instanceError = document.getElementById("instance-error");
-    const name = (instanceInput?.value || "").trim();
-    if (!name) {
-      if (instanceInput) {
-        instanceInput.setAttribute("aria-invalid", "true");
-        instanceInput.focus();
-      }
-      if (instanceError) instanceError.textContent = "Instance name is required";
-      return false;
-    }
-    if (!/^[A-Za-z0-9 _-]+$/.test(name)) {
-      if (instanceInput) {
-        instanceInput.setAttribute("aria-invalid", "true");
-        instanceInput.focus();
-      }
-      if (instanceError) instanceError.textContent = "Instance name can only contain letters, numbers, spaces, underscores, and hyphens";
-      return false;
-    }
-    if (instanceInput) instanceInput.setAttribute("aria-invalid", "false");
-    if (instanceError) instanceError.textContent = "";
-    return true;
-  }
-
-  function syncModalOpenState(ui) {
-    if (ui?.syncModalOpenState) return ui.syncModalOpenState();
-    const open = document.querySelector(".modal.is-open");
-    document.body.classList.toggle("modal-open", !!open);
-  }
-
-  function setHidden(node, hidden) {
-    if (!node) return;
-    node.hidden = hidden;
-    node.classList.toggle("is-hidden", hidden);
-  }
-
-  function buildProgressKey(ctx, config) {
-    if (ctx?.page === "plugin") {
-      const pid = ctx.pluginId || config.pluginId;
-      const inst = ctx.instance || "";
-      return `INKYPI_LAST_PROGRESS:plugin:${pid}:${inst || "_"}`;
-    }
-    return "INKYPI_LAST_PROGRESS";
-  }
-
-  function fadeSkeleton(skel) {
-    if (!skel) return;
-    skel.classList.add('is-hidden');
-    skel.addEventListener('transitionend', () => { skel.style.display = 'none'; }, { once: true });
-  }
-
-  function showInstanceFallback(imgEl, skeleton, fallback) {
-    setHidden(imgEl, true);
-    setHidden(skeleton, true);
-    setHidden(fallback, false);
-  }
-
-  function updateCombinedColorPreview(combined, bgPicker, textPicker) {
-    combined.style.background = bgPicker.value;
-    combined.style.color = textPicker.value;
-  }
-
-  function ensureInlineValidationMessages(result) {
-    if (!result || !Array.isArray(result.invalid)) return;
-    result.invalid.forEach(({ input, message }) => {
-      if (!input) return;
-      const group = input.closest(".form-group") || input.parentElement;
-      if (!group) return;
-      let messageEl = null;
-      const describedByTokens = (input.getAttribute("aria-describedby") || "")
-        .trim()
-        .split(/\s+/)
-        .filter(Boolean);
-      for (const token of describedByTokens) {
-        const candidate = document.getElementById(token);
-        if (candidate?.classList.contains("validation-message")) {
-          messageEl = candidate;
-          break;
-        }
-      }
-      if (!messageEl) {
-        messageEl = document.createElement("span");
-        messageEl.className = "validation-message";
-        messageEl.setAttribute("role", "alert");
-        const baseId = input.id || input.name || "field";
-        messageEl.id = `${baseId}-error`;
-        group.appendChild(messageEl);
-        if (!describedByTokens.includes(messageEl.id)) {
-          describedByTokens.push(messageEl.id);
-        }
-        input.setAttribute("aria-describedby", describedByTokens.join(" "));
-      }
-      input.setAttribute("aria-invalid", "true");
-      messageEl.textContent = message || "This field is invalid";
-      messageEl.style.display = "";
-    });
-  }
-
-  function setCurrentDisplayRefresh(value) {
-    const currTime = document.getElementById("currentDisplayTime");
-    if (!currTime) return;
-    currTime.textContent = value ? new Date(value).toLocaleString() : "—";
-  }
-
-  function initScheduleFormState() {
-    const form = document.getElementById("scheduleForm");
-    if (!form) return;
-    const button = form.querySelector("[data-schedule-submit]");
-    const instanceInput = document.getElementById("instance");
-    const intervalRadio = document.getElementById("refreshTypeInterval");
-    const scheduledRadio = document.getElementById("refreshTypeScheduled");
-    const intervalInput = document.getElementById("scheduleInterval");
-    const unitInput = document.getElementById("scheduleUnit");
-    const timeInput = document.getElementById("scheduleTime");
-    const apiKeyMissing = button?.dataset.apiKeyMissing === "true";
-
-    function isInstanceValid() {
-      const value = (instanceInput?.value || "").trim();
-      return !!value && /^[A-Za-z0-9 _-]+$/.test(value);
-    }
-
-    function isCadenceValid() {
-      if (scheduledRadio?.checked) return !!timeInput?.value;
-      const min = Number(intervalInput?.min || 1);
-      return Number(intervalInput?.value) >= min;
-    }
-
-    function sync() {
-      const scheduled = !!scheduledRadio?.checked;
-      if (intervalInput) intervalInput.disabled = scheduled;
-      if (unitInput) unitInput.disabled = scheduled;
-      if (timeInput) timeInput.disabled = !scheduled;
-      if (!button || apiKeyMissing) return;
-      const valid = isInstanceValid() && isCadenceValid();
-      button.disabled = !valid;
-      button.setAttribute("aria-disabled", valid ? "false" : "true");
-      button.title = valid ? "" : "Complete the schedule fields first";
-    }
-
-    [instanceInput, intervalInput, unitInput, timeInput].forEach((input) => {
-      input?.addEventListener("input", sync);
-      input?.addEventListener("change", sync);
-    });
-    [intervalRadio, scheduledRadio].forEach((input) => {
-      input?.addEventListener("change", sync);
-    });
-    sync();
-  }
-
-  /**
-   * Toggle visibility of the Configure / Style / Schedule sub-panels.
-   * Mirrors the React `tabline` design from the UI handoff (JTN design refresh).
-   * Hoisted to module scope — uses only DOM, no createPluginPage state
-   * (SonarCloud javascript:S7721).
-   */
-  function setPluginSubtab(id) {
-    document.querySelectorAll("[data-plugin-subtab]").forEach((btn) => {
-      const active = btn.dataset.pluginSubtab === id;
-      btn.classList.toggle("active", active);
-      btn.setAttribute("aria-selected", active ? "true" : "false");
-      btn.setAttribute("tabindex", active ? "0" : "-1");
-    });
-    document.querySelectorAll("[data-plugin-subpanel]").forEach((panel) => {
-      const active = panel.dataset.pluginSubpanel === id;
-      panel.hidden = !active;
-      panel.setAttribute("aria-hidden", active ? "false" : "true");
-    });
-  }
+  const {
+    buildProgressKey,
+    ensureInlineValidationMessages,
+    fadeSkeleton,
+    initScheduleFormState,
+    setCurrentDisplayRefresh,
+    setHidden,
+    setPluginSubtab,
+    showInstanceFallback,
+    syncModalOpenState,
+    updateCombinedColorPreview,
+    validateAddToPlaylistAction,
+  } = globalThis.InkyPiPluginPageShared;
+  const { createProgressController } = globalThis.InkyPiPluginPageProgress;
 
   function createPluginPage(config) {
     const ui = globalThis.InkyPiUI || {};
@@ -176,140 +20,11 @@
     const uploadedFiles = (globalThis.uploadedFiles = globalThis.uploadedFiles || {});
     let actionInFlight = false;
     let workflowMode = "configure";
-
-    function saveLastProgressSnapshot(context) {
-      try {
-        const list = document.getElementById("requestProgressList");
-        const textEl = document.getElementById("requestProgressText");
-        const lines = Array.from(list ? list.querySelectorAll("li") : []).map(
-          (li) => {
-            const timeEl = li.querySelector("time");
-            const text = li.textContent || "";
-            const timeText = timeEl ? timeEl.textContent || "" : "";
-            return timeText && text.startsWith(timeText)
-              ? text.slice(timeText.length).trimStart()
-              : text.trim();
-          }
-        );
-        const data = {
-          finishedAtIso: new Date().toISOString(),
-          summary: textEl ? textEl.textContent : "Done",
-          lines,
-          ctx: context || config.progressContext,
-        };
-        localStorage.setItem(buildProgressKey(data.ctx, config), JSON.stringify(data));
-        localStorage.setItem("INKYPI_LAST_PROGRESS", JSON.stringify(data));
-        syncLastProgressButton();
-      } catch (e) { console.warn("Failed to save progress snapshot:", e); }
-    }
-
-    function getLastProgressSnapshot() {
-      const keys = [
-        buildProgressKey(config.progressContext, config),
-        `INKYPI_LAST_PROGRESS:plugin:${config.pluginId}:_`,
-      ];
-      for (const key of keys) {
-        const raw = localStorage.getItem(key);
-        if (!raw) continue;
-        try {
-          return JSON.parse(raw);
-        } catch (e) {
-          console.warn("Corrupt progress data, removing key:", key, e);
-          localStorage.removeItem(key);
-        }
-      }
-      return null;
-    }
-
-    function syncLastProgressButton() {
-      const button = document.getElementById("showLastProgressBtn");
-      if (!button) return;
-      try {
-        const hasSnapshot = !!getLastProgressSnapshot();
-        button.disabled = !hasSnapshot;
-        button.setAttribute("aria-disabled", hasSnapshot ? "false" : "true");
-        button.title = hasSnapshot
-          ? "Show the most recent saved progress log"
-          : "Run Update Preview first to save progress";
-      } catch {
-        button.disabled = true;
-        button.setAttribute("aria-disabled", "true");
-      }
-    }
-
-    function showLastProgress() {
-      try {
-        const data = getLastProgressSnapshot();
-        const progress = document.getElementById("requestProgress");
-        const textEl = document.getElementById("requestProgressText");
-        const clockEl = document.getElementById("requestProgressClock");
-        const elapsedEl = document.getElementById("requestProgressElapsed");
-        const list = document.getElementById("requestProgressList");
-        const bar = document.getElementById("requestProgressBar");
-        // JTN-634: Previously, the no-data path surfaced a toast only, which
-        // users (especially on Weather / AI Image where validation often
-        // blocks the first Update Now attempt before any snapshot is saved)
-        // reported as "no feedback". Show an empty-state inside the progress
-        // block itself so the click always produces a clearly visible result
-        // anchored to the "Last progress" button.
-        if (!data) {
-          if (list) {
-            list.innerHTML = "";
-            const li = document.createElement("li");
-            li.className = "progress-empty-state";
-            li.textContent =
-              "No progress data yet — run Update Now to see progress here.";
-            list.appendChild(li);
-          }
-          if (textEl) textEl.textContent = "No progress data yet";
-          if (clockEl) clockEl.textContent = "—";
-          if (elapsedEl) elapsedEl.textContent = "—";
-          if (bar) {
-            bar.style.width = "0%";
-            const meter = document.getElementById("requestProgressBarMeter");
-            if (meter) meter.value = 0;
-          }
-          if (progress) {
-            setHidden(progress, false);
-            progress.style.display = "";
-          }
-          return;
-        }
-        if (list) {
-          list.innerHTML = "";
-          data.lines.forEach((rawLine) => {
-            const line = rawLine.replace(
-              /^\s*\d{1,2}:\d{2}(?::\d{2})?\s*(AM|PM)?\s*/i,
-              ""
-            );
-            const li = document.createElement("li");
-            const ts = document.createElement("time");
-            ts.textContent = new Date(data.finishedAtIso).toLocaleTimeString();
-            li.appendChild(ts);
-            li.appendChild(document.createTextNode(` ${line}`));
-            list.appendChild(li);
-          });
-        }
-        if (textEl) textEl.textContent = data.summary || "Last run";
-        if (clockEl) {
-          clockEl.textContent = new Date(data.finishedAtIso).toLocaleTimeString();
-        }
-        if (elapsedEl) elapsedEl.textContent = "—";
-        if (bar) {
-          bar.style.width = "100%";
-          const meter = document.getElementById("requestProgressBarMeter");
-          if (meter) meter.value = 100;
-        }
-        // JTN-312: clear the HTML `hidden` attribute so the block is visible;
-        // setting style.display alone does not override the `hidden` attribute.
-        // JTN-347: also clear inline display:none left by progress.stop() —
-        // otherwise the block stays invisible even after removing `hidden`.
-        if (progress) {
-          setHidden(progress, false);
-          progress.style.display = "";
-        }
-      } catch (e) { console.warn("Failed to show last progress:", e); }
-    }
+    const {
+      saveLastProgressSnapshot,
+      showLastProgress,
+      syncLastProgressButton,
+    } = createProgressController({ config, buildProgressKey, setHidden });
 
     function setLatestRefresh(value) {
       if (!value) return;
