+++ tests/integration/test_plugin_draft_add_to_playlist.py
@@ -61,6 +61,8 @@ def test_draft_add_to_playlist_button_reveals_schedule_tab_with_real_handlers(cl
             "src/static/scripts/response_modal.js",
             "src/static/scripts/form_validator.js",
             "src/static/scripts/plugin_form.js",
+            "src/static/scripts/plugin_page/shared.js",
+            "src/static/scripts/plugin_page/progress.js",
             "src/static/scripts/plugin_page.js",
         ):
             with open(path, encoding="utf-8") as f:
@@ -143,6 +145,8 @@ def test_draft_add_to_playlist_click_surfaces_failure_if_schedule_panel_missing(
             "src/static/scripts/response_modal.js",
             "src/static/scripts/form_validator.js",
             "src/static/scripts/plugin_form.js",
+            "src/static/scripts/plugin_page/shared.js",
+            "src/static/scripts/plugin_page/progress.js",
             "src/static/scripts/plugin_page.js",
         ):
             with open(path, encoding="utf-8") as f:
